fix broken warping wip
The problem was that the mesh used for warping initially had a loop in its topology. This didn't seem to hurt when just using the original warper, but broke the warping when the warped output of this was used to generate a new warper (so the warper could be based on the mean mesh)

Using a different original mesh for the warper fixed it.